Forward `HttpRequest`s in `UpstreamHttpService` instead of `Upstream`
